for #2567, split RouteConditionEngine to InsertClauseRouteConditionEngine & WhereClauseRouteConditionEngine
